Fixed checkstyle, pmd and findbugs
