--- PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/graph/nodes/AbstractNode.java
@@ -1,80 +1,0 @@
-package nl.tudelft.pl2016gr2.model.graph.nodes;
-
-import nl.tudelft.pl2016gr2.thirdparty.testing.utility.TestId;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-/**
- * Aides in implementing the <code>Node</code> interface by implementing methods that should have
- * the same behaviour for all nodes.
- * <p>
- * This node only has an ID. it defers all other fields, such as in edges and out edges to its
- * implementations.
- * </p>
- *
- * @author Wouter Smit
- */
-public abstract class AbstractNode implements Node {
-
-  @TestId(id = "id_field")
-  private int identifier;
-
-  /**
-   * Construct a bare abstract node with an ID.
-   *
-   * @param identifier the id of the node.
-   */
-  public AbstractNode(int identifier) {
-    this.identifier = identifier;
-  }
-
-  @Override
-  public int getId() {
-    return identifier;
-  }
-
-  @Override
-  public boolean hasChildren() {
-    return false;
-  }
-  
-  @Override
-  public boolean hasChild(GraphNode child) {
-    return false;
-  }
-
-  @Override
-  public Collection<GraphNode> getChildren() {
-    return null;
-  }
-
-  @Override
-  public Collection<Integer> getGenomesOverEdge(GraphNode node) {
-    //assert getOutEdges().contains(
-    //    node) : "Tried to get genomes over edge for node " + node.getId() + "but it is "
-    //    + "not a direct successor. This = " + this.getId();
-    Collection<Integer> genomes = new ArrayList<>();
-    getGenomes().stream().filter(genome -> node.getGenomes().contains(genome)).forEach(genomes
-        ::add);
-    return genomes;
-  }
-
-  @Override
-  public String toString() {
-    return "id: " + getId();
-  }
-
-  @Override
-  public int hashCode() {
-    return identifier * 37;
-  }
-
-  @Override
-  public boolean equals(Object obj) {
-    if (obj == null || !getClass().equals(obj.getClass())) {
-      return false;
-    }
-    return identifier == ((AbstractNode) obj).identifier;
-  }
-}
